Improved resource documentation surrounding response targets.
Also DRY'd up some of the shared doc logic.
