+++ aws-sdk-core/lib/aws-sdk-core.rb
@@ -111,6 +111,7 @@ module Aws
       autoload :DocstringProvider, 'aws-sdk-core/api/docs/docstring_provider'
       autoload :NullDocstringProvider, 'aws-sdk-core/api/docs/docstring_provider'
       autoload :OperationDocumenter, 'aws-sdk-core/api/docs/operation_documenter'
+      autoload :ParamFormatter, 'aws-sdk-core/api/docs/param_formatter'
       autoload :RequestSyntaxExample, 'aws-sdk-core/api/docs/request_syntax_example'
       autoload :ResponseStructureExample, 'aws-sdk-core/api/docs/response_structure_example'
       autoload :Utils, 'aws-sdk-core/api/docs/utils'
