+++ aws-sdk-resources/lib/aws-sdk-resources/documenter/resource_operation_documenter.rb
@@ -26,17 +26,17 @@ module Aws
         def example_tags
           id = target_resource_class.identifiers.last.to_s
           idv = target_resource_class_name.downcase + '-' + id.gsub('_', '-')
-          tag = []
-          tag << "@example Basic usage"
-          tag << "  #{resp_variable} = #{variable_name}.#{operation_name}(options)"
+          example = []
+          example << "@example Basic usage"
+          example << "  #{resp_variable} = #{variable_name}.#{operation_name}(options)"
           if plural?
-            tag << "  #{resp_variable}.map(&:#{id})"
-            tag << "  #=> [#{idv.inspect}, ...]"
+            example << "  #{resp_variable}.map(&:#{id})"
+            example << "  #=> [#{idv.inspect}, ...]"
           else
-            tag << "  #{resp_variable}.#{id}"
-            tag << "  #=> #{idv.inspect}"
+            example << "  #{resp_variable}.#{id}"
+            example << "  #=> #{idv.inspect}"
           end
-          YARD::DocstringParser.new.parse(tag).to_docstring.tags
+          super + [tag(example.join("\n"))]
         end
 
         def resp_variable
